requirements.txt
PyQt6
pywin32
psutil
Pillow